added improved IDW component which works without pre-calculation of weights
git-svn-id: http://svn.geogr.uni-jena.de/svn/jamsmodels/trunk@1933 503a54e1-1403-0410-a754-803ee147b016

+++ J2K_base/src/org/unijena/j2k/statistics/IDW.java
@@ -56,7 +56,7 @@ public class IDW implements Serializable {
     double pidw;
     Projection proj;
 
-    public void init(double[] statX, double[] statY, double[] statElevation, int pidw, Projection proj) {
+    public void init(double[] statX, double[] statY, double[] statElevation, double pidw, Projection proj) {
         n = statX.length;
         if (dist == null || dist.length != n) {
             dist = new double[n];
